+++ src/Command/PharDownloadCommand.php
@@ -15,10 +15,8 @@ use function sprintf;
  */
 final class PharDownloadCommand implements CommandInterface
 {
-    /** @var string  */
-    private $phar;
-    /** @var string  */
-    private $bin;
+    private string $phar;
+    private string $bin;
 
     /**
      * @param array<string, mixed> $properties