+++ src/Command/PipInstallCommand.php
@@ -15,10 +15,8 @@ use function sprintf;
  */
 final class PipInstallCommand implements CommandInterface
 {
-    /** @var int  */
-    private $pipVersion;
-    /** @var string  */
-    private $requirement;
+    private int $pipVersion;
+    private string $requirement;
 
     /**
      * @param array<string, mixed> $properties