+++ CHANGELOG.md
@@ -11,6 +11,7 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 
 - upgrade [docker-php-extension-installer](https://github.com/mlocati/docker-php-extension-installer) to version 1.4.12
 - `rdfkafka` extension is now supported by PHP 8.1
+- `bin/devkit.php` is made available into the Composer `bin-dir`
 
 ## [1.3.0] - 2022-01-09