+++ CHANGELOG.md
@@ -7,6 +7,8 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 
 ## [Unreleased]
 
+## [1.4.0] - 2022-01-22
+
 ### Added
 
 - `about` command to display current long version and more information about this package.
@@ -267,7 +269,8 @@ Second prototype version with following contents (+40 extensions, +5 tools):
 First prototype version with following contents:
 [ROADMAP](https://github.com/llaville/docker-php-toolbox/blob/e3159c67983107b525270f4770ef8483dd065312/ROADMAP.md)
 
-[unreleased]: https://github.com/llaville/docker-php-toolbox/compare/1.3.0...HEAD
+[unreleased]: https://github.com/llaville/docker-php-toolbox/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/llaville/docker-php-toolbox/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/llaville/docker-php-toolbox/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/llaville/docker-php-toolbox/compare/1.1.1...1.2.0
 [1.1.1]: https://github.com/llaville/docker-php-toolbox/compare/1.1.0...1.1.1